Remove extraneous code in ClampMaker#create_XY_slot_profile_toolpath
Given the recent default settings for slot width and tool diameter, there is no longer a need for a while loop to make a conditional number of machining passes. The method now produces the same slot width every time, and the only slot dimension that changes based on user input is the length.