Reduce spec dependency on implementation
The former spec suite assumed that one set of methods would generate CNC code for the XY plane, and another set would generate CNC code for the Z axis. This was based on a particular implementation where the methods for the Z axis would internally call the methods for the XY plane inside a loop that incremented the Z depth for each iteration, which was tangled-spaghetti-ish.

Now, there are specifications for just three methods (one for generating CNC code for cutting a hole, one for a slot, and one for the outer profile). The implementation of how those methods generate the CNC code is now ready for some major refactoring.