Implement each CNC motion as an individual method
This makes the code more tight and modular by providing lots of methods that do one thing, instead of fewer methods that do lots of things and are confusingly intertwined with each other. It also makes the code more readable by providing individual methods that are named to describe each particular CNC motion.