implement a web server
this commit starts to set up the application for user interaction over the web instead of a command line.

this commit sets up a Rack web server that listens for requests on localhost:8080 and responds with a 200 response code, a header which indicates that the content type is html, and a body with some text in it.