make semantic naming more meaningful
this commit changes a couple of things to make the source code more semantically meaningful. the class ClampMaker is now GCodeGenerator, since the purpose of the class is literally to generate gcode. that class's instance method #generate_gcode is now #generate_clampmaking_gcode to maintain a reference to how the purpose of the gcode is for making a clamp.